+++ src/yw_table_.py
@@ -16,7 +16,7 @@ from pywriter.ui.set_icon_tk import *
 from ywtablelib.ywtable_globals import *
 from ywtablelib.relations_table import RelationsTable
 from ywtablelib.node import Node
-from ywtablelib.scrolled_window import ScrolledWindow
+from ywtablelib.table_frame import TableFrame
 
 APPLICATION = 'Relationship Table'
 APPNAME = 'yw_table'
@@ -33,7 +33,6 @@ class TableManager(MainTk):
 
     def __init__(self, **kwargs):
         super().__init__(f'{APPLICATION}  @release', **kwargs)
-        self.show_status(_('Use the mouse wheel for vertical scrolling, and <Shift>-mouse wheel for horizontal scrolling.'))
         set_icon(self.root, icon='tLogo32')
 
     def open_project(self, fileName):
@@ -43,7 +42,7 @@ class TableManager(MainTk):
         Node.isModified = False
         if self.novel is not None:
             # Set up a window with scrollbars.
-            self._tableWindow = ScrolledWindow(self.mainWindow)
+            self._tableWindow = TableFrame(self.mainWindow)
             self._tableWindow.pack(fill=tk.BOTH, expand=True)
 
             # Build the table structure.