+++ docs/changelog.md
@@ -5,6 +5,13 @@
 ## Changelog
 
 
+### v0.10.0 Beta release
+
+- Make the scrollbars work.
+- Stop scrolling above the canvas window.
+
+Based on PyWriter v9.0.5
+
 ### v0.9.1 Alpha release
 
 - Widen narrow columns.